mypage 디자인패턴 수정
